Created new branch cpp to convert main to c++
The reason for this switch is so that the function to be integrated
can be made a class rather than a plain c function. This way the
function can be constructed with the user supplied values of beta_a
and beta_b. I am hoping that the member function to evaluate the function
at a point can then be passed to and called by the c routine Integrate.